pyproject.toml
[project]
name = "rtvc"
version = "0.1.0"
description = "Real-Time Voice Conversion GUI"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = {text = "MIT"}
keywords = ["voice-conversion", "svc"]
authors = [
    {name = "lengyue", email = "lengyue@lengyue.me"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "noisereduce>=2.0.1",
    "sounddevice>=0.4.6",
    "pyyaml>=6.0",
    "PyQt6>=6.5.0",
    "pyqtdarktheme==2.1.0",
    "requests>=2.31.0",
    "torch>=2.0.1",
]

[project.urls]
repository = "https://github.com/fishaudio/realtime-vc-gui"

[project.gui-scripts]
rtvc = "rtvc.__main__:main"

[tool.pdm]
[tool.pdm.build]
includes = ["rtvc"]

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.3.0",
    "pytest>=7.3.1",
    "PySimpleGUI>=4.60.5",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
lint = { shell = "black . && isort ." }
test = { shell = "PYTHONPATH=. pytest -n=auto -q tests" }
docs = { shell = "sphinx-autobuild docs docs/_build/html" }

[[tool.pdm.source]]
type = "find_links"
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
verify_ssl = true

[tool.isort]
profile = "black"
extend_skip = ["dataset", "logs"]
